Stand 01.12.2020

+++ backend/backend_modifyProduct.php
@@ -1,4 +1,6 @@
 <?php
+    session_start();   
+
     if(!isset($_POST['ID_Produkt']))
     {
         header("Location: http://localhost/Webshop_Melanie_Sebastian/backend/backend_search.php");
@@ -128,7 +130,7 @@
                             <td colspan='2'>
                                 <div class='flex-DirRow flex-SpaceBetween'> 
                                     <input type='submit' value='Speichern' class='button buttonNormal'></input>
-                                    <a href='deleteProduct.php?productToDelete=".$_POST['ID_Produkt']."' class='button buttonNormal'>Löschen</a>
+                                    <a href='deleteProduct.php?productToDelete=".$_POST['ID_Produkt']."' class='button buttonNormal' onclick='return confirm(\"Wollen sie dieses Produkt wirklich löschen?\");'>Löschen</a>
                                 </div>
                             </td>
                         </tr>
